+++ test/karma/expression.spec.js
@@ -158,4 +158,14 @@ describe('ExpresssionProcessor', function () {
     });
   });
 
+  describe('_queryCache', function() {
+    it('should be unique for each ExpressionProcessor', function() {
+      var lfData = new LForms.LFormsData({fhirVersion: 'R4', items: []});
+      var exp1 = new LForms.FHIR.R4.SDC.ExpressionProcessor(lfData);
+      exp1._queryCache[1] = 2;
+      var exp2 = new LForms.FHIR.R4.SDC.ExpressionProcessor(lfData);
+      assert.equal(Object.keys(exp2._queryCache).length, 0);
+      assert.equal(Object.keys(exp1._queryCache).length, 1);
+    });
+  });
 });